qml: set the Setting control's implicitWidth to 450
In the onboarding views that represent advanced settings customization,
we essentially set the max width to a certain value so that the settings
components stretch to the correct max value of 450. If we are to use the
OnboardingInfo control for the settings views, we will no longer have
an independent Header control stretching the parent's width and thus the
settings component will no longer be drawn to a max width of 450.
Setting an implicit width of 450 for the Settings control fixes this
issue.